# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SANOVIM — Marketing Médico com IA API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /video/process:
    post:
      operationId: processVideo
      tags: [video]
      summary: Trim, caption, and convert video to Reels format
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/ProcessVideoRequest"
      responses:
        "200":
          description: Processed video URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessedVideo"
  /video/history:
    get:
      operationId: getVideoHistory
      tags: [video]
      summary: Video processing history
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Video history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VideoHistoryItem"

  # ── Market Predictions ────────────────────────────────────────────────
...
components:
  schemas:
...
    ProcessVideoRequest:
      type: object
      properties:
        startSeconds:
          type: number
          default: 0
        endSeconds:
          type: number
        addCaptions:
          type: boolean
          default: false
        convertToReels:
          type: boolean
          default: true
        video:
          type: string
          format: binary
      required: [video]
    ProcessedVideo:
      type: object
      properties:
        id:
          type: integer
        outputUrl:
          type: string
        duration:
          type: number
        format:
          type: string
        captions:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, outputUrl, duration, createdAt]
    VideoHistoryItem:
      type: object
      properties:
        id:
          type: integer
        originalName:
          type: string
        outputUrl:
          type: string
        duration:
          type: number
        hasCaptions:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, outputUrl, duration, createdAt]
...